Create covid_19_data.ipynb
This Jupyter Notebook will import the dataset into a Pandas DataFrame so it can be manipulated in Python. This will let users get up-and-running using this data with just a couple clicks. It's also implemented in Google Colab, which allows users to run it on a virtual machine through their browsers, with no need to install any kernel or packages locally.

I think this would be a handy tool to include with the data, and please let me know if you have any questions or would like anything changed. If you like this idea, I'm happy to add things to the notebook like interactive visualizations and analyses.